Update Ex_03_C2.py
fixes on the script so it can run float values instead of only integers

+++ Ex_03_C2.py
@@ -1,4 +1,4 @@
-h=input('Enter hours:')
-r=input('Enter rate:')
-p=int(h)*int(r)
+h=input('Enter hours:', )
+r=input('Enter rate:', )
+p=float(h)*float(r)
 print('Your pay is:', p)
